Move all parseSQL utils into utils file
